GeoTrace_Forensics/requirements.txt
PyQt5
Pillow
exifread
folium
reportlab
openpyxl
pandas